+++ server/deva-ts/src/App.tsx
@@ -48,5 +48,4 @@ function App() {
 }
 
 
-
 export default App;
